fix bug that keep slow log on, dble may occur oom (#2640)
