+++ src/main/java/com/actiontech/dble/server/handler/ServerLoadDataInfileHandler.java
@@ -218,6 +218,7 @@ public final class ServerLoadDataInfileHandler implements LoadDataInfileHandler 
                     if (rrs != null) {
                         flushDataToFile();
                         isStartLoadData = false;
+                        serverConnection.getSession2().endParse();
                         serverConnection.getSession2().execute(rrs);
                     }
                 }
@@ -646,6 +647,7 @@ public final class ServerLoadDataInfileHandler implements LoadDataInfileHandler 
         RouteResultset rrs = buildResultSet(routeResultMap);
         if (rrs != null) {
             flushDataToFile();
+            serverConnection.getSession2().endParse();
             serverConnection.getSession2().execute(rrs);
         }
     }
